+++ src/main/java/dev/aura/bungeechat/listener/MOTDListener.java
@@ -1,16 +1,12 @@
 package dev.aura.bungeechat.listener;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import dev.aura.bungeechat.account.BungeecordAccountManager;
 import dev.aura.bungeechat.api.account.BungeeChatAccount;
 import dev.aura.bungeechat.api.enums.Permission;
 import dev.aura.bungeechat.api.placeholder.BungeeChatContext;
 import dev.aura.bungeechat.event.BungeeChatJoinEvent;
+import dev.aura.bungeechat.message.Format;
 import dev.aura.bungeechat.message.MessagesService;
-import dev.aura.bungeechat.message.PlaceHolderUtil;
-import dev.aura.bungeechat.module.BungeecordModuleManager;
 import dev.aura.bungeechat.permission.PermissionManager;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Listener;
@@ -25,10 +21,7 @@ public class MOTDListener implements Listener {
             return;
 
         BungeeChatAccount bungeeChatAccount = BungeecordAccountManager.getAccount(player).get();
-        List<String> motd = BungeecordModuleManager.MOTD_MODULE.getModuleSection().getStringList("message");
 
-        String message = motd.stream().collect(Collectors.joining("\n"));
-        MessagesService.sendMessage(player,
-                PlaceHolderUtil.formatMessage(message, new BungeeChatContext(bungeeChatAccount)));
+        MessagesService.sendMessage(player, Format.MOTD.get(new BungeeChatContext(bungeeChatAccount)));
     }
 }